Package now relocatable
Qt compiles in the installation path. If you move the installation, qmake can no longer locate various files. The prefix path can be overriden with qt.conf.

Previously, the recipe put the onus on the package consumer to create a qt.conf file. It was unclear to the consumer that this should be done and it meant the Qt installation could not be used stand-alone or with QtCreator without jumping through hoops.

The package step now creates a qt.conf in the bin directory, thus making the package relocatable.